core/requirements.txt
amqp==5.0.5
asgiref==3.3.1
billiard==3.6.3.0
celery==5.0.5
click==7.1.2
click-didyoumean==0.0.3
click-plugins==1.1.1
click-repl==0.1.6
django==3.1.7
djangorestframework==3.12.2
gunicorn==20.0.4
kombu==5.0.2
pillow==8.1.2
psycopg2-binary==2.8
prompt-toolkit==3.0.17
pytz==2021.1
six==1.15.0
sqlparse==0.4.1
vine==5.0.0
wcwidth==0.2.5
numpy==1.19.5
keras==2.4.3
